reducing resoluion of projection plots

+++ Ramses_analysis/FU_ori_investigation/apastron_frames.py
@@ -65,7 +65,7 @@ end_time = plot_times[-1]
 plt.clf()
 fig = plt.figure(figsize=(two_col_width, 0.8*two_col_width))
 fig.subplots_adjust(left=0, right=1, bottom=0, top=1)
-G = gridspec.GridSpec(3,n_frames/2)#,height_ratios=[1.5, 2, 2])
+G = gridspec.GridSpec(3,int(n_frames/2))#,height_ratios=[1.5, 2, 2])
 axes_1 = plt.subplot(G[0, :])
 plt.subplots_adjust(wspace=0.0)
 #plt.subplots_adjust(wspace=-0.4)
